flask_backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dateutil==2.8.2
pymongo==4.5.0
bcrypt==4.0.1
